requirements.txt
Django==5.0
Pillow==10.1.0
factory-boy==3.3.0
enum34==1.1.10
pycountry==23.12.11